various additions
